User reviews save to database correctly

+++ src/main/resources/db/migration/V5__create_user_reviews_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE user_reviews (
   id BIGSERIAL PRIMARY KEY,
   title VARCHAR(50) NOT NULL,
   content VARCHAR(1000) NOT NULL,
-  rating SMALLINT NOT NULL,
+  rating INT NOT NULL,
   timestamp TIMESTAMP,
   userid BIGINT REFERENCES users(id),
   authorid BIGINT REFERENCES users(id)
